# internal/httpapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AIV2API",
    "version": "1.5.0",
    "description": "Image, video, and audio generation API. Video, audio, and asynchronous image endpoints return tasks.",
    "license": {"name": "All rights reserved", "identifier": "LicenseRef-Proprietary"}
  },
  "servers": [{"url": "/", "description": "Current host"}],
...
  "paths": {
...
    "/v1/models": {
      "get": {"tags": ["Models"], "operationId": "listModels", "summary": "List models allowed to this API key", "responses": {"200": {"description": "Curated public models allowed to this API key", "content": {"application/json": {"schema": {"type": "object", "properties": {"object": {"const": "list"}, "data": {"type": "array", "items": {"$ref": "#/components/schemas/Model"}}}}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}}}
    },
    "/v1/images/estimate": {
      "post": {
        "tags": ["Images"],
        "operationId": "estimateImageCost",
        "summary": "Estimate image credits without submitting a generation",
        "description": "Validates image parameters and returns the credits that a generation request would reserve. It does not create a task.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ImageEstimateRequest"}}}},
        "responses": {
          "200": {"description": "Exact image credit estimate", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ImageEstimate"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "422": {"$ref": "#/components/responses/CostUnavailable"}
        }
      }
    },
    "/v1/images/generations": {
      "post": {
        "tags": ["Images"],
        "operationId": "createImage",
        "summary": "Generate images",
        "description": "Synchronous OpenAI Images-compatible response. Idempotency-Key is required. When the synchronous wait budget expires, the API returns 202 with the durable task and a Location header instead of creating a second operation.",
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ImageRequest"}}}},
        "responses": {
          "200": {"description": "Generated images", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ImageResult"}}}},
          "202": {"description": "Synchronous wait expired; continue with the returned task and Location header", "headers": {"Location": {"schema": {"type": "string"}}, "Retry-After": {"schema": {"type": "integer"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "402": {"$ref": "#/components/responses/InsufficientBalance"}, "409": {"$ref": "#/components/responses/IdempotencyConflict"}, "422": {"$ref": "#/components/responses/ImageConversionFailed"}, "429": {"$ref": "#/components/responses/RateLimited"}, "502": {"$ref": "#/components/responses/UpstreamImageFailure"}, "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/v1/images/edits": {
      "post": {
        "tags": ["Images"], "operationId": "editImage", "summary": "Edit images using references", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/ImageEditRequest"}}}},
        "responses": {"200": {"description": "Edited images", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ImageResult"}}}}, "202": {"description": "Synchronous wait expired; continue polling the returned task", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "402": {"$ref": "#/components/responses/InsufficientBalance"}, "409": {"$ref": "#/components/responses/IdempotencyConflict"}, "413": {"description": "Reference image exceeds the configured size limit", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}, "415": {"description": "Reference image type is not PNG, JPEG, or WebP", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}, "422": {"$ref": "#/components/responses/ImageConversionFailed"}, "429": {"$ref": "#/components/responses/RateLimited"}, "502": {"$ref": "#/components/responses/UpstreamImageFailure"}, "503": {"$ref": "#/components/responses/ServiceUnavailable"}}
      }
    },
    "/v1/tasks/images": {
      "post": {
        "tags": ["Images", "Tasks"], "operationId": "createAsyncImageTask", "summary": "Create an asynchronous image task", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AsyncImageRequest"}}}},
        "responses": {"202": {"description": "Task created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "200": {"description": "Existing idempotent task", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "402": {"$ref": "#/components/responses/InsufficientBalance"}, "409": {"$ref": "#/components/responses/IdempotencyConflict"}, "422": {"$ref": "#/components/responses/CostUnavailable"}, "429": {"$ref": "#/components/responses/RateLimited"}, "503": {"$ref": "#/components/responses/ServiceUnavailable"}}
      }
    },
    "/v1/videos/generations": {
      "post": {
        "tags": ["Videos"],
        "operationId": "createVideo",
        "summary": "Create a video task",
        "description": "The five entries below are request templates, not values for a mode field and not all mutually exclusive. The API has no mode field: it infers reference inputs from uploaded multipart fields. No-reference requests use application/json; requests with reference media use multipart/form-data. Seedance can combine image, video, and audio references; Kling O3 Omni accepts up to seven images, start/end frames, or one video plus up to four images; FLUX 3 Video supports either start/end frames or one video continuation reference; MiniMax H3 can combine image and audio references. Frame inputs are exclusive with all other reference media. Grok Imagine 1.5 is image-to-video only: start_frame is required and end_frame is unsupported.",
        "x-generation-modes": [
          {"id": "none", "name": "No reference", "content_type": "application/json", "models": ["flux-3-video", "seedance-2.0", "seedance-2.0-fast", "seedance-2.0-mini", "veo-3.1", "veo-3.1-fast", "kling-o3-omni", "minimax-h3"], "required_media": [], "rule": "Do not send image, start_frame, end_frame, video, or audio."},
          {"id": "image", "name": "Image reference", "content_type": "multipart/form-data", "models": ["seedance-2.0", "seedance-2.0-fast", "seedance-2.0-mini", "veo-3.1", "kling-o3-omni", "minimax-h3"], "required_media": ["image or image[]"], "rule": "Cannot be combined with start_frame or end_frame. Veo 3.1 fixes size=1280x720 and duration=8. Kling O3 Omni accepts seven images, or four when a video reference is also present."},
          {"id": "frame", "name": "Start/end frame", "content_type": "multipart/form-data", "models": ["flux-3-video", "seedance-2.0", "seedance-2.0-fast", "seedance-2.0-mini", "veo-3.1", "veo-3.1-fast", "kling-o3-omni", "minimax-h3", "grok-imagine-1.5"], "required_media": ["start_frame"], "optional_media": ["end_frame"], "model_exceptions": {"grok-imagine-1.5": "start_frame is required and end_frame is unsupported"}, "rule": "end_frame requires start_frame. Frames cannot be combined with ordinary images, videos, or audio. Grok Imagine 1.5 requires start_frame and rejects end_frame."},
          {"id": "video", "name": "Video reference", "content_type": "multipart/form-data", "models": ["flux-3-video", "seedance-2.0", "seedance-2.0-fast", "seedance-2.0-mini", "kling-o3-omni"], "required_media": ["video or video[]"], "rule": "Kling O3 Omni accepts one 3-10.05 second file and up to four images; duration must match the file within 0.05 seconds and 2160p is unavailable. FLUX 3 Video accepts one file up to 50 MB and 15.05 seconds; Seedance 2.0 models accept 3 files totaling 15 seconds. Cannot be combined with start_frame or end_frame."},
          {"id": "audio", "name": "Audio reference", "content_type": "multipart/form-data", "models": ["seedance-2.0", "seedance-2.0-fast", "seedance-2.0-mini", "minimax-h3"], "required_media": ["audio or audio[]", "supporting reference"], "rule": "Seedance requires an image or video and accepts one audio file totaling 15 seconds. MiniMax H3 requires an ordinary image and accepts three totaling 15 seconds."}
        ],
        "x-reference-combinations": {
          "mode_field": false,
          "inferred_from": "multipart file fields",
          "frame_exclusive_with": ["image", "video", "audio"],
          "end_frame_requires": "start_frame",
          "generate_audio_scope": "output_native_audio",
          "rules": [
            {
              "models": ["seedance-2.0", "seedance-2.0-fast", "seedance-2.0-mini"],
              "combinable_reference_fields": ["image", "video", "audio"],
              "audio_requires_any": ["image", "video"]
            },
            {
              "models": ["flux-3-video"],
              "combinable_reference_fields": ["video"],
              "alternative_reference_fields": ["start_frame", "end_frame"],
              "unsupported_reference_fields": ["image", "audio"]
            },
            {
              "models": ["kling-o3-omni"],
              "combinable_reference_fields": ["image", "video"],
              "video_reference_max_images": 4,
              "alternative_reference_fields": ["start_frame", "end_frame"],
              "unsupported_reference_fields": ["audio"]
            },
            {
              "models": ["minimax-h3"],
              "combinable_reference_fields": ["image", "audio"],
              "audio_requires_all": ["image"]
            },
            {
              "models": ["grok-imagine-1.5"],
              "combinable_reference_fields": ["start_frame"],
              "required_reference_fields": ["start_frame"],
              "unsupported_reference_fields": ["image", "end_frame", "video", "audio"]
            }
          ]
        },
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VideoRequest"}}, "multipart/form-data": {"schema": {"$ref": "#/components/schemas/VideoReferenceRequest"}}}},
        "responses": {"202": {"description": "Task created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "200": {"description": "Existing idempotent task", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "402": {"$ref": "#/components/responses/InsufficientBalance"}, "409": {"$ref": "#/components/responses/IdempotencyConflict"}, "422": {"$ref": "#/components/responses/CostUnavailable"}, "429": {"$ref": "#/components/responses/RateLimited"}, "503": {"$ref": "#/components/responses/ServiceUnavailable"}}
      }
    },
    "/v1/videos/estimate": {
      "post": {
        "tags": ["Videos"],
        "operationId": "estimateVideoCost",
        "summary": "Estimate video credits without submitting a generation",
        "description": "Validates video parameters and returns the credits that a task would reserve. It does not create a task.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VideoEstimateRequest"}}}},
        "responses": {
          "200": {"description": "Exact video credit estimate", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/VideoEstimate"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "422": {"$ref": "#/components/responses/CostUnavailable"}
        }
      }
    },
    "/v1/audio/generations": {
      "post": {
        "tags": ["Audio"], "operationId": "createAudio", "summary": "Create an audio task", "description": "Supports text-to-speech, music, and sound effects.", "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AudioRequest"}}}},
        "responses": {"202": {"description": "Task created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "200": {"description": "Existing idempotent task", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "402": {"$ref": "#/components/responses/InsufficientBalance"}, "409": {"$ref": "#/components/responses/IdempotencyConflict"}, "422": {"$ref": "#/components/responses/CostUnavailable"}, "429": {"$ref": "#/components/responses/RateLimited"}, "503": {"$ref": "#/components/responses/ServiceUnavailable"}}
      }
    },
    "/v1/chat/completions": {
      "post": {
        "tags": ["Chat", "Images"], "operationId": "createChatCompletion", "summary": "Create an image through Chat Completions compatibility",
        "description": "Uses the final user message as the image prompt. Accepts text and one Base64 image. Set stream=true for SSE.",
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatCompletionRequest"}}}},
        "responses": {"200": {"description": "Chat completion or SSE stream", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatCompletion"}}, "text/event-stream": {"schema": {"type": "string"}}}}, "202": {"description": "Synchronous wait expired; continue polling the returned task", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "402": {"$ref": "#/components/responses/InsufficientBalance"}, "409": {"$ref": "#/components/responses/IdempotencyConflict"}, "422": {"$ref": "#/components/responses/CostUnavailable"}, "429": {"$ref": "#/components/responses/RateLimited"}, "502": {"$ref": "#/components/responses/UpstreamImageFailure"}, "503": {"$ref": "#/components/responses/ServiceUnavailable"}}
      }
    },
    "/v1/tasks/{id}": {"get": {"tags": ["Tasks"], "operationId": "getTask", "summary": "Get a task", "parameters": [{"$ref": "#/components/parameters/TaskID"}], "responses": {"200": {"description": "Task state and result", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/v1/tasks/{id}/cancel": {"post": {"tags": ["Tasks"], "operationId": "cancelTask", "summary": "Cancel a queued task", "parameters": [{"$ref": "#/components/parameters/TaskID"}], "responses": {"200": {"description": "Task cancelled", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CancelResult"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}, "409": {"description": "Only queued tasks can be cancelled", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}}}},
    "/v1/videos/{id}": {"get": {"tags": ["Videos", "Tasks"], "operationId": "getVideoTask", "summary": "Get a video task", "parameters": [{"$ref": "#/components/parameters/TaskID"}], "responses": {"200": {"description": "Task state and video result", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/v1/videos/{id}/cancel": {"post": {"tags": ["Videos", "Tasks"], "operationId": "cancelVideoTask", "summary": "Cancel a queued video task", "parameters": [{"$ref": "#/components/parameters/TaskID"}], "responses": {"200": {"description": "Task cancelled", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CancelResult"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}, "409": {"description": "Task is no longer cancellable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}}}},
    "/v1/audio/{id}": {"get": {"tags": ["Audio", "Tasks"], "operationId": "getAudioTask", "summary": "Get an audio task", "parameters": [{"$ref": "#/components/parameters/TaskID"}], "responses": {"200": {"description": "Task state and audio result", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}}}},
    "/v1/audio/{id}/cancel": {"post": {"tags": ["Audio", "Tasks"], "operationId": "cancelAudioTask", "summary": "Cancel a queued audio task", "parameters": [{"$ref": "#/components/parameters/TaskID"}], "responses": {"200": {"description": "Task cancelled", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CancelResult"}}}}, "400": {"$ref": "#/components/responses/BadRequest"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}, "409": {"description": "Task is no longer cancellable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}}}}
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "API key"}},
    "parameters": {
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "description": "Required unique value for one paid business operation. Reuse it only when retrying the identical request body. A missing value returns idempotency_key_required; a value over 255 characters returns idempotency_key_too_long.", "schema": {"type": "string", "minLength": 1, "maxLength": 255}},
...
    },
    "schemas": {
...
      "Model": {"type": "object", "required": ["id", "object", "created", "owned_by"], "additionalProperties": false, "properties": {"id": {"type": "string"}, "object": {"const": "model"}, "created": {"type": "integer", "format": "int64"}, "owned_by": {"const": "aiv2api"}}},
...
      "ImageRequest": {
        "type": "object", "required": ["model", "prompt"], "additionalProperties": false,
        "description": "Strict public image DTO. Leonardo-native prompt_enhance/style_ids, internal references, and Nano Banana 2 video_reference_base are not exposed and are rejected as unknown fields.",
        "properties": {
          "model": {"type": "string", "enum": ["gpt-image-2", "nano-banana-2", "nano-banana-pro", "seedream-5.0-pro"]},
          "prompt": {"type": "string", "minLength": 1, "maxLength": 9999},
          "size": {"$ref": "#/components/schemas/ImageSize"},
          "n": {"$ref": "#/components/schemas/ImageQuantity"},
          "quality": {"type": "string", "enum": ["auto", "low", "medium", "high"], "default": "low", "description": "gpt-image-2 only. auto is normalized to low by this gateway; Leonardo's native schema default is MEDIUM."},
          "response_format": {"type": "string", "enum": ["url", "b64_json"]},
          "output_format": {"type": "string", "enum": ["png", "jpeg"], "description": "Used with response_format=b64_json"},
          "output_compression": {"type": "integer", "minimum": 0, "maximum": 100, "description": "JPEG only"},
          "background": {"type": "string", "enum": ["auto", "opaque"]},
          "moderation": {"const": "auto"}
        }
      },
      "ImageEstimateRequest": {
        "type": "object", "required": ["model"], "additionalProperties": false,
        "description": "Image pricing inputs only. size accepts auto or a model-valid WIDTHxHEIGHT; quality is accepted only for gpt-image-2; n is the requested image quantity.",
        "examples": [{"model": "gpt-image-2", "size": "1536x1024", "quality": "high", "n": 1}],
        "properties": {
          "model": {"type": "string", "enum": ["gpt-image-2", "nano-banana-2", "nano-banana-pro", "seedream-5.0-pro"]},
          "size": {"$ref": "#/components/schemas/ImageSize"},
          "quality": {"type": "string", "enum": ["auto", "low", "medium", "high"], "default": "auto", "description": "gpt-image-2 only; auto is normalized to low. Other models reject quality."},
          "n": {"$ref": "#/components/schemas/ImageQuantity"}
        }
      },
      "ImageEstimate": {
        "type": "object", "required": ["model", "size", "quantity", "unit_tokens", "estimated_tokens", "pricing_basis", "formula", "cost_parameters", "price_version", "source"],
        "additionalProperties": false,
        "examples": [{"model": "gpt-image-2", "size": "1536x1024", "quality": "high", "quantity": 1, "unit_tokens": 388, "estimated_tokens": 388, "pricing_basis": "pixel_formula", "pricing_anchor": "2048x2048", "quality_multiplier": 35.167, "formula": "ceil(width * height / 1000000 * 7 * quality_multiplier) * n", "cost_parameters": ["model", "size", "quality", "n"], "price_version": "2026-07-15-ui", "source": "leonardo-ui"}],
        "properties": {
          "model": {"type": "string"},
          "size": {"type": "string", "description": "Normalized exact WIDTHxHEIGHT used for the estimate."},
          "quality": {"type": "string", "description": "Normalized GPT quality; omitted for models without a quality parameter."},
          "quantity": {"type": "integer", "minimum": 1},
          "unit_tokens": {"type": "integer", "minimum": 0, "description": "Credits for one output."},
          "estimated_tokens": {"type": "integer", "minimum": 0, "description": "unit_tokens multiplied by quantity; this is the reservation amount."},
          "pricing_basis": {"type": "string", "enum": ["pixel_formula", "size_tier", "size_threshold", "fixed"]},
          "pricing_tier": {"type": "string", "enum": ["small", "medium", "large", "standard", "2k"]},
          "pricing_anchor": {"type": "string", "description": "Internal pricing bucket used to select the active rule."},
          "quality_multiplier": {"type": "number", "description": "GPT Image 2 multiplier: low=1, medium=8.833, high=35.167."},
          "formula": {"type": "string"},
          "cost_parameters": {"type": "array", "items": {"type": "string"}, "description": "Request fields that can change the estimate for this model."},
          "price_version": {"type": "string"},
          "source": {"type": "string"}
        }
      },
...
    },
    "responses": {
      "BadRequest": {"description": "Invalid request parameters", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "examples": {"invalid_request": {"value": {"error": {"message": "prompt is required", "type": "invalid_request", "code": "invalid_request"}}}, "idempotency_key_required": {"value": {"error": {"message": "Idempotency-Key header is required", "type": "idempotency_key_required", "code": "idempotency_key_required"}}}, "idempotency_key_too_long": {"value": {"error": {"message": "Idempotency-Key must not exceed 255 characters", "type": "idempotency_key_too_long", "code": "idempotency_key_too_long"}}}}}}},
      "Unauthorized": {"description": "Missing or invalid Bearer API key", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "example": {"error": {"message": "missing Bearer API key", "type": "invalid_api_key", "code": "invalid_api_key"}}}}},
...
      "InsufficientBalance": {"description": "Account pool balance is below the estimated generation cost", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "example": {"error": {"message": "the account pool does not have enough available credits", "type": "insufficient_pool_balance", "code": "insufficient_pool_balance"}}}}},
...
      "CostUnavailable": {"description": "No active price rule matches the request parameters", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}, "example": {"error": {"message": "cost is unavailable for the selected model parameters", "type": "cost_unavailable", "code": "cost_unavailable"}}}}},
...
    }
  },
  "security": [{"bearerAuth": []}]
}